v85-added editor-uploads route for handling richtext editor images . its not completed for now

+++ back-end/src/helpers/deleteWrapper.js
@@ -4,7 +4,9 @@ const deleteFile = require("./deleteFile");
 
 const deleteWrapper = (fileURLS) => {
   //if some files uploaded with this req , delete them
-  for (let urlObj of Object.values(fileURLS)) 
-    deleteFile(urlObj.url.substring(1));
+  for (let urlObj of Object.values(fileURLS))
+    if (urlObj) {
+      deleteFile(urlObj.url.substring(1));
+    }
 };
 module.exports = deleteWrapper;
